Finally.  Revamped input.
Had to screw around with streams to get the stupid first character that
you typed to not show up in the messages, and I'm not entirely sure that
I even had to.  But I'm gonna go with that I did.

Ended up being a relatively simple implementation, so that's good.  This
is theoretically bug free, but I'm sure brendan will find a way to break
it.  Let the testing commence.